Fix resolving external refs to platform/role docs on Hydra
platformDocObjectsInventory now expects a file path to an objects.inv
generated by the platform/role docs.

(To clarify: this repo here is also called "platform docs" sometimes,
not to be confused with the branch-specific platform/role docs from the "doc" dir
in the flyingcircusio/fc-nixos repo).

The inventory path is normally passed in by Hydra as a local Nix store path, fetched
from a Hydra input (type "local path or URL") pointing to the current
build product of the job doc.roles from the default platform version, for example:
https://hydra.flyingcircus.io/job/flyingcircus/fc-21.11-dev/doc.roles/latest/download-by-type/file/inventory