Fix auto theme selection
Using the Python customization points offered by the Furo theme
instead of CSS fixes the automatic light/dark mode selection.

There are some design differences at the moment, though and testing
changes is a bit harder using the Python way.

I inlined the logos to make it work but that means that we must not
forget updating it here, too. Probably remote loading is better but
it's not directly supported by Furo, I think.