Copy webgateway docs to fix ref from older platform versions
The kubernetes/k3s doc starting with 22.05 refers to `webgateway` which
is interpreted by intersphinx as a reference to the Gentoo webgateway
doc. We already fixed that for 24.05 but the old versions still refer
to it. It's easier to vendor the wanted doc here than to update all the
outdated versions.